assets/MinecraftServerManager/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
mctools==1.3.0
python-dotenv==1.0.1
flask-cors==5.0.1